Modify M236 and setOutputPressure to use floating point number instead of integer for desired pressure.

+++ Marlin/Regulator.cpp
@@ -8,8 +8,8 @@
 
 
 // Set the output pressure (in psi) of the regulator.
-void setOutputPressure(uint8_t pressure) {
-	// 4095/(130.355 psi)
+void setOutputPressure(float pressure) {
+	// 4095/(130.355 psi) (adjusted from theoretical value for better accuracy)
 	uint16_t digital_val = (uint16_t)(BITS_PER_PSI * pressure);
 	// Write value to DAC
 	DAC_write(MCP4725_I2C_ADDRESS, digital_val);
